requirements.txt
playwright>=1.50.0
pytest>=8.0.0
python-dotenv>=1.0.0
